requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
stripe==7.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.1
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3